fix: Correct syntax errors in job recovery mechanism
