+++ cli/src/client/contract.ts
@@ -48,7 +48,7 @@ export const VersionedTextsSchema = z.object({
     z.object({
       version: z.string(),
       content: z.string(),
-    }),
+    })
   ),
 });
 
@@ -87,17 +87,13 @@ export const resultDataSchema = z
 export const learningSchema = z.object({
   summary: z
     .string()
-    .describe(
-      "The new information that was learned. Be generic, do not refer to the entities.",
-    ),
+    .describe("The new information that was learned. Be generic, do not refer to the entities."),
   entities: z
     .array(
       z.object({
-        name: z
-          .string()
-          .describe("The name of the entity this learning relates to."),
+        name: z.string().describe("The name of the entity this learning relates to."),
         type: z.enum(["tool"]),
-      }),
+      })
     )
     .describe("The entities this learning relates to."),
   relevance: z.object({
@@ -121,7 +117,7 @@ export const agentDataSchema = z
           toolName: z.string(),
           reasoning: z.string().optional(),
           input: z.object({}).passthrough(),
-        }),
+        })
       )
       .optional(),
   })
@@ -162,7 +158,7 @@ export const definition = {
             description: z.string().optional(),
             schema: z.string().optional(),
             config: FunctionConfigSchema.optional(),
-          }),
+          })
         )
         .optional(),
     }),
@@ -226,13 +222,11 @@ export const definition = {
         message: z.string(),
       }),
     },
-    body: blobSchema
-      .omit({ id: true, createdAt: true, jobId: true, workflowId: true })
-      .and(
-        z.object({
-          data: z.string(),
-        }),
-      ),
+    body: blobSchema.omit({ id: true, createdAt: true, jobId: true, workflowId: true }).and(
+      z.object({
+        data: z.string(),
+      })
+    ),
   },
   getContract: {
     method: "GET",
@@ -256,7 +250,7 @@ export const definition = {
           name: z.string(),
           createdAt: z.date(),
           description: z.string().nullable(),
-        }),
+        })
       ),
       401: z.undefined(),
     },
@@ -271,9 +265,7 @@ export const definition = {
       204: z.undefined(),
     },
     body: z.object({
-      description: z
-        .string()
-        .describe("Human readable description of the cluster"),
+      description: z.string().describe("Human readable description of the cluster"),
     }),
   },
   upsertIntegrations: {
@@ -332,13 +324,13 @@ export const definition = {
       name: z.string().optional(),
       description: z.string().optional(),
       additionalContext: VersionedTextsSchema.optional().describe(
-        "Additional cluster context which is included in all runs",
+        "Additional cluster context which is included in all runs"
       ),
       debug: z
         .boolean()
         .optional()
         .describe(
-          "Enable additional logging (Including prompts and results) for use by Inferable support",
+          "Enable additional logging (Including prompts and results) for use by Inferable support"
         ),
       enableRunConfigs: z.boolean().optional(),
       enableKnowledgebase: z.boolean().optional(),
@@ -389,7 +381,7 @@ export const definition = {
           workflowId: z.string().nullable(),
           meta: z.any().nullable(),
           id: z.string(),
-        }),
+        })
       ),
       401: z.undefined(),
       404: z.undefined(),
@@ -437,10 +429,7 @@ export const definition = {
         .string()
         .optional()
         .describe("An initial 'human' message to trigger the run"),
-      systemPrompt: z
-        .string()
-        .optional()
-        .describe("A system prompt for the run."),
+      systemPrompt: z.string().optional().describe("A system prompt for the run."),
       name: z
         .string()
         .optional()
@@ -452,24 +441,20 @@ export const definition = {
       resultSchema: anyObject
         .optional()
         .describe(
-          "A JSON schema definition which the result object should conform to. By default the result will be a JSON object which does not conform to any schema",
+          "A JSON schema definition which the result object should conform to. By default the result will be a JSON object which does not conform to any schema"
         ),
       attachedFunctions: z
         .array(functionReference)
         .optional()
         .describe(
-          "An array of functions to make available to the run. By default all functions in the cluster will be available",
+          "An array of functions to make available to the run. By default all functions in the cluster will be available"
         ),
       onStatusChange: z
         .object({
-          function: functionReference.describe(
-            "A function to call when the run status changes",
-          ),
+          function: functionReference.describe("A function to call when the run status changes"),
         })
         .optional()
-        .describe(
-          "Mechanism for receiving notifications when the run status changes",
-        ),
+        .describe("Mechanism for receiving notifications when the run status changes"),
       metadata: z
         .record(z.string())
         .optional()
@@ -480,21 +465,16 @@ export const definition = {
           mocks: z
             .record(
               z.object({
-                output: z
-                  .object({})
-                  .passthrough()
-                  .describe("The mock output of the function"),
-              }),
+                output: z.object({}).passthrough().describe("The mock output of the function"),
+              })
             )
             .optional()
             .describe(
-              "Function mocks to be used in the run. (Keys should be function in the format <SERVICE>_<FUNCTION>)",
+              "Function mocks to be used in the run. (Keys should be function in the format <SERVICE>_<FUNCTION>)"
             ),
         })
         .optional()
-        .describe(
-          "When provided, the run will be marked as as a test / evaluation",
-        ),
+        .describe("When provided, the run will be marked as as a test / evaluation"),
       config: z
         .object({
           id: z.string().describe("The run configuration ID"),
@@ -502,7 +482,7 @@ export const definition = {
             .object({})
             .passthrough()
             .describe(
-              "The run configuration input arguments, the schema must match the run configuration input schema",
+              "The run configuration input arguments, the schema must match the run configuration input schema"
             )
             .optional(),
         })
@@ -517,11 +497,7 @@ export const definition = {
         })
         .optional()
         .describe("DEPRECATED"),
-      reasoningTraces: z
-        .boolean()
-        .default(true)
-        .optional()
-        .describe("Enable reasoning traces"),
+      reasoningTraces: z.boolean().default(true).optional().describe("Enable reasoning traces"),
       callSummarization: z
         .boolean()
         .default(true)
@@ -530,9 +506,7 @@ export const definition = {
       interactive: z
         .boolean()
         .default(true)
-        .describe(
-          "Allow the run to be continued with follow-up messages / message edits",
-        ),
+        .describe("Allow the run to be continued with follow-up messages / message edits"),
     }),
     responses: {
       201: z.object({
@@ -609,7 +583,7 @@ export const definition = {
           createdAt: z.date(),
           pending: z.boolean().default(false),
           displayableContext: z.record(z.string()).nullable(),
-        }),
+        })
       ),
       401: z.undefined(),
     },
@@ -624,13 +598,10 @@ export const definition = {
       userId: z.string().optional(),
       test: z.coerce
         .string()
-        .transform((value) => value === "true")
+        .transform(value => value === "true")
         .optional(),
       limit: z.coerce.number().min(10).max(50).default(50),
-      metadata: z
-        .string()
-        .optional()
-        .describe("Filter runs by a metadata value (value:key)"),
+      metadata: z.string().optional().describe("Filter runs by a metadata value (value:key)"),
       configId: z.string().optional(),
     }),
     responses: {
@@ -640,14 +611,12 @@ export const definition = {
           name: z.string(),
           userId: z.string().nullable(),
           createdAt: z.date(),
-          status: z
-            .enum(["pending", "running", "paused", "done", "failed"])
-            .nullable(),
+          status: z.enum(["pending", "running", "paused", "done", "failed"]).nullable(),
           test: z.boolean(),
           configId: z.string().nullable(),
           configVersion: z.number().nullable(),
           feedbackScore: z.number().nullable(),
-        }),
+        })
       ),
       401: z.undefined(),
     },
@@ -662,9 +631,7 @@ export const definition = {
       200: z.object({
         id: z.string(),
         userId: z.string().nullable(),
-        status: z
-          .enum(["pending", "running", "paused", "done", "failed"])
-          .nullable(),
+        status: z.enum(["pending", "running", "paused", "done", "failed"]).nullable(),
         failureReason: z.string().nullable(),
         test: z.boolean(),
         feedbackComment: z.string().nullable(),
@@ -684,12 +651,7 @@ export const definition = {
     }),
     body: z.object({
       comment: z.string().describe("Feedback comment").nullable(),
-      score: z
-        .number()
-        .describe("Score between 0 and 1")
-        .min(0)
-        .max(1)
-        .nullable(),
+      score: z.number().describe("Score between 0 and 1").min(0).max(1).nullable(),
     }),
     responses: {
       204: z.undefined(),
@@ -811,7 +773,7 @@ export const definition = {
           targetFn: z.string(),
           service: z.string(),
           executingMachineId: z.string().nullable(),
-        }),
+        })
       ),
     },
   },
@@ -847,7 +809,7 @@ export const definition = {
           createdAt: z.date(),
           createdBy: z.string(),
           revokedAt: z.date().nullable(),
-        }),
+        })
       ),
     },
   },
@@ -877,7 +839,7 @@ export const definition = {
           id: z.string(),
           lastPingAt: z.date(),
           ip: z.string(),
-        }),
+        })
       ),
     },
     pathParams: z.object({
@@ -903,10 +865,10 @@ export const definition = {
                 description: z.string().optional(),
                 schema: z.string().optional(),
                 config: FunctionConfigSchema.optional(),
-              }),
+              })
             )
             .optional(),
-        }),
+        })
       ),
     },
     pathParams: z.object({
@@ -945,7 +907,7 @@ export const definition = {
             createdAt: z.date(),
             pending: z.boolean().default(false),
             displayableContext: z.record(z.string()).nullable(),
-          }),
+          })
         ),
         activity: z.array(
           z.object({
@@ -956,7 +918,7 @@ export const definition = {
             createdAt: z.date(),
             jobId: z.string().nullable(),
             targetFn: z.string().nullable(),
-          }),
+          })
         ),
         jobs: z.array(
           z.object({
@@ -968,14 +930,12 @@ export const definition = {
             createdAt: z.date(),
             approved: z.boolean().nullable(),
             approvalRequested: z.boolean().nullable(),
-          }),
+          })
         ),
         run: z.object({
           id: z.string(),
           userId: z.string().nullable(),
-          status: z
-            .enum(["pending", "running", "paused", "done", "failed"])
-            .nullable(),
+          status: z.enum(["pending", "running", "paused", "done", "failed"]).nullable(),
           failureReason: z.string().nullable(),
           test: z.boolean(),
           feedbackComment: z.string().nullable(),
@@ -1082,7 +1042,7 @@ export const definition = {
             resultSchema: anyObject.nullable(),
             inputSchema: anyObject.nullable(),
             public: z.boolean(),
-          }),
+          })
         ),
       }),
       401: z.undefined(),
@@ -1103,10 +1063,7 @@ export const definition = {
     body: z.object({
       name: z.string(),
       initialPrompt: z.string().optional(),
-      systemPrompt: z
-        .string()
-        .optional()
-        .describe("The initial system prompt for the run."),
+      systemPrompt: z.string().optional().describe("The initial system prompt for the run."),
       attachedFunctions: z.array(z.string()).optional(),
       resultSchema: anyObject.optional(),
       inputSchema: z.object({}).passthrough().optional().nullable(),
@@ -1131,10 +1088,7 @@ export const definition = {
     body: z.object({
       name: z.string().optional(),
       initialPrompt: z.string().optional(),
-      systemPrompt: z
-        .string()
-        .optional()
-        .describe("The initial system prompt for the run."),
+      systemPrompt: z.string().optional().describe("The initial system prompt for the run."),
       attachedFunctions: z.array(z.string()).optional(),
       resultSchema: z.object({}).passthrough().optional().nullable(),
       inputSchema: z.object({}).passthrough().optional().nullable(),
@@ -1187,7 +1141,7 @@ export const definition = {
           resultSchema: z.unknown().nullable(),
           createdAt: z.date(),
           updatedAt: z.date(),
-        }),
+        })
       ),
       401: z.undefined(),
     },
@@ -1214,7 +1168,7 @@ export const definition = {
           createdAt: z.date(),
           updatedAt: z.date(),
           similarity: z.number(),
-        }),
+        })
       ),
       401: z.undefined(),
     },
@@ -1234,7 +1188,7 @@ export const definition = {
           jobFailureCount: z.number(),
           timeToCompletion: z.number(),
           jobCount: z.number(),
-        }),
+        })
       ),
     },
     pathParams: z.object({
@@ -1251,11 +1205,9 @@ export const definition = {
         z.object({
           id: z.string(),
           data: z.string(),
-          tags: z
-            .array(z.string())
-            .transform((tags) => tags.map((tag) => tag.toLowerCase().trim())),
+          tags: z.array(z.string()).transform(tags => tags.map(tag => tag.toLowerCase().trim())),
           title: z.string(),
-        }),
+        })
       ),
     }),
     responses: {
@@ -1283,7 +1235,7 @@ export const definition = {
           tags: z.array(z.string()),
           title: z.string(),
           similarity: z.number(),
-        }),
+        })
       ),
       401: z.undefined(),
     },
@@ -1297,9 +1249,7 @@ export const definition = {
     headers: z.object({ authorization: z.string() }),
     body: z.object({
       data: z.string(),
-      tags: z
-        .array(z.string())
-        .transform((tags) => tags.map((tag) => tag.toLowerCase().trim())),
+      tags: z.array(z.string()).transform(tags => tags.map(tag => tag.toLowerCase().trim())),
       title: z.string(),
     }),
     responses: {
@@ -1334,7 +1284,7 @@ export const definition = {
           data: z.string(),
           tags: z.array(z.string()),
           title: z.string(),
-        }),
+        })
       ),
       401: z.undefined(),
     },
@@ -1363,9 +1313,7 @@ export const definition = {
         .min(0)
         .max(20)
         .default(0)
-        .describe(
-          "Time in seconds to keep the request open waiting for a response",
-        ),
+        .describe("Time in seconds to keep the request open waiting for a response"),
     }),
     headers: z.object({
       authorization: z.string(),
@@ -1413,14 +1361,9 @@ export const definition = {
     path: "/clusters/:clusterId/calls",
     query: z.object({
       service: z.string(),
-      status: z
-        .enum(["pending", "running", "paused", "done", "failed"])
-        .default("pending"),
+      status: z.enum(["pending", "running", "paused", "done", "failed"]).default("pending"),
       limit: z.coerce.number().min(1).max(20).default(10),
-      acknowledge: z.coerce
-        .boolean()
-        .default(false)
-        .describe("Should calls be marked as running"),
+      acknowledge: z.coerce.boolean().default(false).describe("Should calls be marked as running"),
     }),
     pathParams: z.object({
       clusterId: z.string(),
@@ -1442,7 +1385,7 @@ export const definition = {
           authContext: z.any().nullable(),
           runContext: z.any().nullable(),
           approved: z.boolean(),
-        }),
+        })
       ),
     },
   },
@@ -1474,7 +1417,7 @@ export const definition = {
       resultSchema: anyObject
         .optional()
         .describe(
-          "A JSON schema definition which the result object should conform to. By default the result will be a JSON object which does not conform to any schema",
+          "A JSON schema definition which the result object should conform to. By default the result will be a JSON object which does not conform to any schema"
         ),
       modelId: z.enum(["claude-3-5-sonnet", "claude-3-haiku"]),
       temperature: z
