+++ control-plane/drizzle/0203_sad_speed.sql
@@ -1,0 +1,2 @@
+ALTER TABLE "jobs" ALTER COLUMN "remaining_attempts" SET DEFAULT 1;--> statement-breakpoint
+ALTER TABLE "machines" DROP COLUMN IF EXISTS "status";
